+++ beacon-chain/blockchain/BUILD.bazel
@@ -5,6 +5,7 @@ go_library(
     srcs = [
         "fork_choice.go",
         "service.go",
+        "stateGen.go",
     ],
     importpath = "github.com/prysmaticlabs/prysm/beacon-chain/blockchain",
     visibility = ["//beacon-chain:__subpackages__"],
@@ -30,6 +31,7 @@ go_test(
     srcs = [
         "fork_choice_test.go",
         "service_test.go",
+        "stateGen_test.go",
     ],
     embed = [":go_default_library"],
     deps = [
