+++ beacon-chain/blockchain/stateGen/stateGen_test.go
@@ -1,0 +1,57 @@
+package stateGen
+
+import (
+	"testing"
+
+	"github.com/prysmaticlabs/prysm/beacon-chain/chaintest/backend"
+	"github.com/prysmaticlabs/prysm/beacon-chain/db"
+)
+
+func TestGenerateState_OK(t *testing.T) {
+	bd, err := backend.NewSimulatedBackend()
+	if err != nil {
+		t.Fatalf("Could not create a new simulated backend %v", err)
+	}
+	privKeys, err := bd.SetupBackend(100)
+	if err != nil {
+		t.Fatalf("Could not set up backend %v", err)
+	}
+	beaconDb := bd.DB()
+	defer bd.Shutdown()
+	defer db.TeardownDB(beaconDb)
+
+	slotLimit := uint64(30)
+
+	// Run the simulated chain for 30 slots, to get a state that we can save as finalized
+	for i := uint64(0); i < slotLimit; i++ {
+		if err := bd.GenerateBlockAndAdvanceChain(&backend.SimulatedObjects{}, privKeys); err != nil {
+			t.Fatalf("Could not generate block and transition state successfully %v for slot %d", err, bd.State().Slot+1)
+		}
+	}
+
+	if err := beaconDb.SaveFinalizedState(bd.State()); err != nil {
+		t.Fatalf("Unable to save finalized state")
+	}
+
+	// Run the chain for another 30 slots so that we can have this at the current head.
+	for i := uint64(0); i < slotLimit; i++ {
+		if err := bd.GenerateBlockAndAdvanceChain(&backend.SimulatedObjects{}, privKeys); err != nil {
+			t.Fatalf("Could not generate block and transition state successfully %v for slot %d", err, bd.State().Slot+1)
+		}
+	}
+
+	// Save all in memory blocks
+
+	for _, v := range bd.InMemoryBlocks() {
+		beaconDb.SaveBlock(v)
+	}
+
+	if err := beaconDb.SaveState(bd.State()); err != nil {
+		t.Fatalf("Unable to save state in chain %v", err)
+	}
+
+	_, err = GenerateStateFromSlot(beaconDb, 60)
+	if err != nil {
+		t.Fatalf("Unable to generate new state from previous finalized state %v", err)
+	}
+}
